#86 :truck: 상세 화면 서브 뷰 페이저 프래그먼트들 폴더링함
