+++ app/src/main/java/io/github/seoj17/canyongg/ui/detail/analysisTab/pages/damaged/TeamDamagedViewHolder.kt
@@ -1,4 +1,4 @@
-package io.github.seoj17.canyongg.ui.detail.analysisTab.pages.viewHolder
+package io.github.seoj17.canyongg.ui.detail.analysisTab.pages.damaged
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -7,22 +7,22 @@ import io.github.seoj17.canyongg.databinding.ItemTeamAnalysisBinding
 import io.github.seoj17.canyongg.ui.model.SummonerMatchRecord
 import io.github.seoj17.canyongg.utils.setChampion
 
-class TeamDealtViewHolder(
+class TeamDamagedViewHolder(
     private val binding: ItemTeamAnalysisBinding,
 ) : RecyclerView.ViewHolder(binding.root) {
     fun bind(data: SummonerMatchRecord, maxValue: Int) {
         with(binding) {
             champion.setChampion(data.championName)
             summonerName.text = data.summonerName
-            itemValue.text = data.totalDealt.toString()
-            valueGraph.progress = data.totalDealt
+            itemValue.text = data.totalDamaged.toString()
+            valueGraph.progress = data.totalDamaged
             valueGraph.max = maxValue
         }
     }
 
     companion object {
-        operator fun invoke(parent: ViewGroup): TeamDealtViewHolder {
-            return TeamDealtViewHolder(
+        operator fun invoke(parent: ViewGroup): TeamDamagedViewHolder {
+            return TeamDamagedViewHolder(
                 ItemTeamAnalysisBinding.inflate(
                     LayoutInflater.from(parent.context), parent, false
                 )
