+++ app/src/main/java/io/github/seoj17/canyongg/ui/detail/analysisTab/pages/gold/TeamGoldFragment.kt
@@ -1,4 +1,4 @@
-package io.github.seoj17.canyongg.ui.detail.analysisTab.pages
+package io.github.seoj17.canyongg.ui.detail.analysisTab.pages.gold
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,12 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import dagger.hilt.android.AndroidEntryPoint
-import io.github.seoj17.canyongg.databinding.FragmentTeamVisionScoreBinding
+import io.github.seoj17.canyongg.databinding.FragmentTeamGoldBinding
+import io.github.seoj17.canyongg.ui.detail.analysisTab.pages.AnalysisPageListAdapter
 
 @AndroidEntryPoint
-class TeamVisionScoreFragment : Fragment() {
-    private lateinit var binding: FragmentTeamVisionScoreBinding
-    private val viewModel: TeamVisionScoreViewModel by viewModels()
+class TeamGoldFragment : Fragment() {
+    private lateinit var binding: FragmentTeamGoldBinding
+    private val viewModel: TeamGoldViewModel by viewModels()
 
     override fun onCreateView(
         inflater: LayoutInflater,
@@ -20,7 +21,7 @@ class TeamVisionScoreFragment : Fragment() {
         savedInstanceState: Bundle?,
     ): View {
         super.onCreateView(inflater, container, savedInstanceState)
-        binding = FragmentTeamVisionScoreBinding.inflate(layoutInflater, container, false)
+        binding = FragmentTeamGoldBinding.inflate(layoutInflater, container, false)
         return binding.root
     }
 
@@ -29,13 +30,13 @@ class TeamVisionScoreFragment : Fragment() {
         with(binding) {
             lifecycleOwner = viewLifecycleOwner
             vm = viewModel
-            teamVisionScoreList.adapter = AnalysisPageListAdapter(4)
+            teamGoldList.adapter = AnalysisPageListAdapter(2)
         }
     }
 
     companion object {
-        fun newInstance(matchId: String): TeamVisionScoreFragment {
-            return TeamVisionScoreFragment().apply {
+        fun newInstance(matchId: String): TeamGoldFragment {
+            return TeamGoldFragment().apply {
                 arguments = Bundle().apply {
                     putString("matchId", matchId)
                 }
