+++ app/src/main/java/io/github/seoj17/canyongg/ui/detail/analysisTab/pages/kill/TeamKillsViewHolder.kt
@@ -1,4 +1,4 @@
-package io.github.seoj17.canyongg.ui.detail.analysisTab.pages.viewHolder
+package io.github.seoj17.canyongg.ui.detail.analysisTab.pages.kill
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -7,22 +7,22 @@ import io.github.seoj17.canyongg.databinding.ItemTeamAnalysisBinding
 import io.github.seoj17.canyongg.ui.model.SummonerMatchRecord
 import io.github.seoj17.canyongg.utils.setChampion
 
-class TeamGoldViewHolder(
+class TeamKillsViewHolder(
     private val binding: ItemTeamAnalysisBinding,
 ) : RecyclerView.ViewHolder(binding.root) {
     fun bind(data: SummonerMatchRecord, maxValue: Int) {
         with(binding) {
             champion.setChampion(data.championName)
             summonerName.text = data.summonerName
-            itemValue.text = data.spentGold.toString()
-            valueGraph.progress = data.spentGold
+            itemValue.text = data.kill.toString()
+            valueGraph.progress = data.kill
             valueGraph.max = maxValue
         }
     }
 
     companion object {
-        operator fun invoke(parent: ViewGroup): TeamGoldViewHolder {
-            return TeamGoldViewHolder(
+        operator fun invoke(parent: ViewGroup): TeamKillsViewHolder {
+            return TeamKillsViewHolder(
                 ItemTeamAnalysisBinding.inflate(
                     LayoutInflater.from(parent.context), parent, false
                 )
