+++ app/src/main/res/layout/fragment_team_damaged.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="vm"
-            type="io.github.seoj17.canyongg.ui.detail.analysisTab.pages.TeamDamagedViewModel" />
+            type="io.github.seoj17.canyongg.ui.detail.analysisTab.pages.damaged.TeamDamagedViewModel" />
 
         <import type="android.view.View" />
     </data>
