+++ app/src/main/res/layout/fragment_team_dealt.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="vm"
-            type="io.github.seoj17.canyongg.ui.detail.analysisTab.pages.TeamDealtViewModel" />
+            type="io.github.seoj17.canyongg.ui.detail.analysisTab.pages.dealt.TeamDealtViewModel" />
 
         <import type="android.view.View" />
     </data>
