+++ app/src/main/res/layout/fragment_team_gold.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="vm"
-            type="io.github.seoj17.canyongg.ui.detail.analysisTab.pages.TeamGoldViewModel" />
+            type="io.github.seoj17.canyongg.ui.detail.analysisTab.pages.gold.TeamGoldViewModel" />
 
         <import type="android.view.View" />
     </data>
