+++ app/src/main/res/layout/fragment_team_kills.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="vm"
-            type="io.github.seoj17.canyongg.ui.detail.analysisTab.pages.TeamKillsViewModel" />
+            type="io.github.seoj17.canyongg.ui.detail.analysisTab.pages.kill.TeamKillsViewModel" />
 
         <import type="android.view.View" />
     </data>
