+++ app/src/main/res/layout/fragment_team_vision_score.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="vm"
-            type="io.github.seoj17.canyongg.ui.detail.analysisTab.pages.TeamVisionScoreViewModel" />
+            type="io.github.seoj17.canyongg.ui.detail.analysisTab.pages.vision.TeamVisionScoreViewModel" />
 
         <import type="android.view.View" />
     </data>
